+++ lib/src/chart/line_chart/line_chart_data.dart
@@ -1040,10 +1040,7 @@ class LineTouchTooltipData with EquatableMixin {
   /// [LineChart] shows a tooltip popup on top of spots automatically when touch happens,
   /// otherwise you can show it manually using [LineChartData.showingTooltipIndicators].
   /// Tooltip shows on top of rods, with [getTooltipColor] as a background color.
-  /// You can set the corner radius using [tooltipRoundedRadius],
-  /// or if you need a custom border, you can use [tooltipBorderRadius].
-  /// Note that if both [tooltipRoundedRadius] and [tooltipBorderRadius] are set,
-  /// the value from [tooltipBorderRadius] will be used.
+  /// You can set the corner radius using [tooltipBorderRadius],
   /// If you want to have a padding inside the tooltip, fill [tooltipPadding],
   /// or If you want to have a bottom margin, set [tooltipMargin].
   /// Content of the tooltip will provide using [getTooltipItems] callback, you can override it
@@ -1053,7 +1050,6 @@ class LineTouchTooltipData with EquatableMixin {
   /// you can set [fitInsideHorizontally] true to force it to shift inside the chart horizontally,
   /// also you can set [fitInsideVertically] true to force it to shift inside the chart vertically.
   const LineTouchTooltipData({
-    double? tooltipRoundedRadius = 4,
     BorderRadius? tooltipBorderRadius,
     this.tooltipPadding =
         const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
@@ -1068,23 +1064,14 @@ class LineTouchTooltipData with EquatableMixin {
     this.showOnTopOfTheChartBoxArea = false,
     this.rotateAngle = 0.0,
     this.tooltipBorder = BorderSide.none,
-  })  :
-        // TODO(imaNNeo): We should remove this property in the next major version
-        // ignore: deprecated_member_use_from_same_package
-        tooltipRoundedRadius = tooltipRoundedRadius ?? 4,
-        _tooltipBorderRadius = tooltipBorderRadius;
+  }) : _tooltipBorderRadius = tooltipBorderRadius;
 
   /// Sets a rounded radius for the tooltip.
-  @Deprecated('use tooltipBorderRadius instead')
-  final double tooltipRoundedRadius;
-
   final BorderRadius? _tooltipBorderRadius;
 
   /// Sets a rounded radius for the tooltip.
   BorderRadius get tooltipBorderRadius =>
-      // TODO(imaNNeo): We should remove this property in the next major version
-      // ignore: deprecated_member_use_from_same_package
-      _tooltipBorderRadius ?? BorderRadius.circular(tooltipRoundedRadius);
+      _tooltipBorderRadius ?? BorderRadius.circular(4);
 
   /// Applies a padding for showing contents inside the tooltip.
   final EdgeInsets tooltipPadding;
@@ -1125,9 +1112,6 @@ class LineTouchTooltipData with EquatableMixin {
   /// Used for equality check, see [EquatableMixin].
   @override
   List<Object?> get props => [
-        // TODO(imaNNeo): We should remove this property in the next major version
-        // ignore: deprecated_member_use_from_same_package
-        tooltipRoundedRadius,
         _tooltipBorderRadius,
         tooltipPadding,
         tooltipMargin,